Distribution versioning and checksum (#101)
## What is the goal of this PR?

### Adds version number to the distribution
Files that are deployed to Github used to look like this: `grakn-core-all-linux.tar.gz` and when extracted will be located in a directory `grakn-core-all-linux`.

With this PR they will look like this: `grakn-core-all-linux-1.5.0.tar.gz` and when extracted will be located in a directory `grakn-core-all-linux-1.5.0`.

## What are the changes implemented in this PR?

A new rule `assemble_versioned` is created. It outputs a distribution which contains the version number. Additionally, `checksum`, and `deploy_github` takes input from `assemble_versioned` rather than `assemble_zip` and `assemble_tar`