Fix an issue where assemble-maven and deploy-maven would incorrectly use project name and description as group and artifact id (#136)
## What is the goal of this PR?

This PR fixes an issue where assemble-maven and deploy-maven would incorrectly use project name and description as group and artifact id. This is caused by the fact that the process of parsing `groupId`, `artifactId`, and `version` relied on brittle array indexing where we're expecting them to be element number 1, 2,and 3 in the `pom.xml`.

This logic has been made robust by retrieving the element by name, ie., `root.find('groupId')`.