Pass tags to pkg_tar (#168)
## What is the goal of this PR?

To pass the `tags` property to pkg_tar. The motivation is that I want to signal to our CI runner to skip this in normal presubmit builds.

```python
tags = ["skipci"],
```

Then this target wouldn't be built as part CI. 

## What are the changes implemented in this PR?

This PR passes the `tags` property to pkg_tar as part of assemble_targz.