Updated workspace refs to include tags and added distribution_file (#238)
## What is the goal of this PR?

To support `deploy_distribution` better, we need to be able to retrieve the version from getting it as a dependency (done with `workspace_refs`). To do this, we have added the ability for workspace_refs to detect the commit or tag of a dependency from bazel tags (so that http_file is supported). We have also created a macro to make creating the `http_file` dependency more explicit and handle adding this tag for us.

## What are the changes implemented in this PR?

- Updated `workspace_refs` to also use tags.
- Added `distribution_file` workspace dependency macro.