Allow java_deps to target Maven packages and support overrides for the output path (#264)
## What is the goal of this PR?

To allow Grakn 2.0 to output rocksdb-dev to lib/dev and rocksdb-prod to lib/prod when building a Grakn distribution.

## What are the changes implemented in this PR?

Firstly, we need to be able to target Maven packages using `java_deps`. We add support for this by looking through both `ctx.attr.target.data_runfiles.files` and `ctx.attr.target.files`.

We also add support for `java_deps` to have output path overrides. When a JAR matching the given pattern is found, it is deployed to the folder specified in the override, instead of the default output folder.