Make the pom_replace_version.py script a py_binary (#260)
## What is the goal of this PR?

Make the rule more hermetic since it currently depends of the presence of
Python in the PATH. With this patch, it would use the interpreter from the Python toolchain.

## What are the changes implemented in this PR?

Wrap the `pom_replace_version` script in a `py_binary`. Note the `assemble` script is already built as a `py_binary`.