Adds a parameter in deploy_github that allows for creating both a draft as well as a published release (#269)
## What is the goal of this PR?

We have added a `draft` parameter to `deploy_github`. If it is set to `True`, the rule will create a release that is marked as draft. When set to `False`, the release will be immediately published.

This is an improvement of the existing behaviour, in which you can only create a draft release.

## What are the changes implemented in this PR?

- Implement the `draft` field
- Refactor the rule and `deploy.py` template