Make java_deps fail on seeing duplicate Maven coordinates #339
## What is the goal of this PR?

Make `java_deps` fail if it finds duplicate Maven coordinates.

## What are the changes implemented in this PR?

We recently had a situation in TypeDB Studio where two distinct `kt_jvm_library` targets passed to `java_deps` had the same Maven coordinates, due to a copy-paste error.

Now, `java_deps` actually does check for duplicates and filter out duplicates. We assume the intent of this is to ensure that a JAR included by two distinct dependencies is only included once in the output bundle. However, this is not legitimate and should never happen. So we are no longer going to do that - instead, we will throw.

The new logic is:

For each JAR file, not part of the JDK, that `ctx.attr.target` imports as a runtime dependency:
1. Check that it has a Maven coordinates tag, if this `java_deps` mandates them
2. Determine the output path of this JAR. If it has Maven coordinates, replace `.` and `:` in them to `-`.
3. Add the output path and JAR to a map (of output paths to JARs). If the output filename was already taken by another JAR, throw. The values of this map are the primary outputs of this `java_deps`.
4. Determine the _full_ output path of this JAR, normally `java_deps_root + output_path + ".jar"`, but may be overridden using the `output_path_overrides` map (e.g: to enable some JARs to be included on the classpath only when certain criteria are met). This is used to output a "JAR mapping file" which can be used as a reference.