Explicitly use python3 (#358)
## What is the goal of this PR?

We now use the command `python3` explicitly instead of `python`. As all
of our code is written in Python 3 and the command `python` is ambiguous
for historical reasons, this is more in line with our current approach
in using java and bazel. That is, we rely upon the image that is
provided to specify the appropriate version of these binaries rather
than doing so ourselves.

## What are the changes implemented in this PR?

We have changed every line specifying 
```
#!/usr/bin/env python
```
to
```
#!/usr/bin/env python3
```