[assemble_maven] add developers tag support for pom (to support Maven Central) (#348)
## What is the goal of this PR?

Support publishing to Maven Central again. Was originally done in #133,
but the `developers` tag was never migrated to the maven rule rewrite in
#296.

## What are the changes implemented in this PR?

Add an optional `developers` attribute to `assemble_maven` rule as a
`string_list_dict`. Developers can be described as such:

```
developers = {
  "sugarmanz": ["name=Jeremiah Zucker", "email=zucker.jeremiah@gmail.com"],
  "adierkens": ["name=Adam Dierkens"],
}
```

And will be populated as:

```xml
<developers>
  <developer>
    <id>adierkens</id>
    <name>Adam Dierkens</name>
  </developer>
  <developer>
    <id>sugarmanz</id>
    <name>Jeremiah Zucker</name>
    <email>zucker.jeremiah@gmail.com</email>
  </developer>
</developers>
```