Upgrade `bazel-skylib` to `1.4.0` and regenerate `README` (#369)
## What is the goal of this PR?

We upgraded `bazel-skylib` to `1.4.0` and regenerated the `README` file.

## What are the changes implemented in this PR?

Recently GitHub was affected by a major incident where commit-based
archives changed their SHAs, crippling our whole build infrastructure.
`bazel-skylib` was one such affected package.

As a precaution, we've upgraded it to the latest version. This has also
resulted in some incidental but helpful bugfixes in our generated
`README.md`, which we've committed.

Also, the Packer OSX SHA256 hash was out of date, presumably due to
Hashicorp regenerating their distribution bundle. We've fixed it and
verified it against their provided document of expected hashes.