Fix assemble_pip and deploy_pip (#368)
Currently there are some issues with these two rules: 'assemble_pip' and
'deploy_pip'.

assemble_pip:
(1) data files are not included in "tarball" build;
(2) it cannot produce "wheel" file which is a common needs for many use
cases.

deploy_pip:
(1) currently the deployment relies on "repository_registry" and
username&password. However, we use "~/.pypirc" to config twine and we
believe many other users have the same config. "pypi_profile" attr is
added to enable this use case;
(2) this rule will run into an error due to missing deps in
requirements.txt;
(3) this rule will run into an error due to that the "repo_type_key" arg
in deploy.py is missing when running the script. Practically, only one
of "release" and "snapshot" can be used for deployment;
(4) this rule can only deploy single package ("tar.gz") which is
incompatible with what we did for "assemble_pip".

This PR is to address all of the issues above.