Fix maven pre-deploy snapshot version validation: require SemVer-SHA (#398)
## What is the goal of this PR?

Fix regression introduced in #397: prepending `0.0.0-` to the commit SHA
for snapshot assembly broke the validation in the deployment script,
which expects either bare SHA or any string ending with `-SNAPSHOT`. We
now require either a major.minor.patch version with the SHA at the end,
or any string ending with `-SNAPSHOT`.