Add Bazel rules for C# packaging, distribution and documentation. (#406)
## What is the goal of this PR?

We prepare tools for the C# driver release, namely: documentation,
packaging, and distribution.

## What are the changes implemented in this PR?

Added:
Rules for the C# nuget packages generation and distribution. These rules
are based on the rules from [the open source Selenium
repo](https://github.com/SeleniumHQ/selenium/), so thanks to them for
boosting the development by making it available! They are used and
tested in the [typedb-driver](https://github.com/vaticle/typedb-driver/)
project:
* `nuget_pack`: gets a list of C# libraries, native libs files (if
needed), a template nuget project description, a list of files to
include into the package (for example, README or logos), and produces
two files: nupkg (nuget package) and snupkg (symbol package for
debugging).
* `nuget_push`: receives the result of the previous command and pushes
the packages (with their symbols) to the target (snapshot/release) repo.

Updated:
* doxygen configuration, previously used solely for `C++`. Now it's also
convenient for `C#`, so some refactoring and additions happened.