Improve Apple code signing (#419)
## Usage and product changes

- We remove the `skipIfSigned` functionality from the `AppleCodeSigner`
- We add verbose logging to `xcrun` calls if it is enabled

## Motivation

We've concluded that it is not the responsibility of a maintainer of an
open-source library to sign their code for MacOS installers, as such we
want to sign these packages ourselves. This may cause us to assume
responsibility for issues in our application caused by these packages;
this seems reasonable.

We have previously encountered issues when releasing `typedb-studio` due
to a package that was signed, but not specifically for MacOS installers,
failing Apple's checks. As such, we want to sign *all* packages captured
by our `deepSignJarsRegex` - not just unsigned ones. This ensures that
all signatures are valid.

However, we still value the `deepSignJarsRegex` to avoid unnecessarily
unpackaging and re-signing every single jar we depend on.

## Implementation

We remove the `skipIfSigned` argument and the functionality gated behind
it.

We also fix an issue where the `verbose` logging flag was not being
correctly propagated to our code signing commands.